installer: tftp server may not be dhcp default.
This configuration could happen when chaining network boot to a
different server. /etc/dhcp-prefix in this case will not be relevant to
the new given tftp server.
In general it seems using /etc/dhcp-prefix is relevant when accessing
the tftp server given by the DHCP server (option 210?). This is obvious
when accessing tftp://dhcp/<path>.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>